+++ tests/proptests/main.rs
@@ -13,3 +13,4 @@
 // limitations under the License.
 
 mod proptest_parse;
+mod proptest_policy;